Populate tableView from DB works ! Hurray !
