--- src/main/java/sample/dto/ProductDTO.java
@@ -1,24 +1,0 @@
-package sample.dto;
-
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-import lombok.ToString;
-
-@Getter
-@Setter
-@ToString
-@NoArgsConstructor
-public class ProductDTO {
-    private String name;
-    private int price;
-    private int quantity;
-    private String description;
-
-    public ProductDTO(String name, int price, int quantity, String description) {
-        this.name = name;
-        this.price = price;
-        this.quantity = quantity;
-        this.description = description;
-    }
-}
