+++ src/main/resources/scene/welcome.fxml
@@ -8,16 +8,16 @@
 <?import javafx.scene.layout.BorderPane?>
 <?import javafx.scene.layout.HBox?>
 
-<BorderPane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="400.0" prefWidth="600.0" xmlns="http://javafx.com/javafx/11.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="sample.controller.Controller">
+<BorderPane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="400.0" prefWidth="600.0" xmlns="http://javafx.com/javafx/11.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="sample.controller.WelcomeController">
    <top>
       <HBox alignment="TOP_RIGHT" prefHeight="400.0" prefWidth="613.0" spacing="20.0" BorderPane.alignment="CENTER">
          <children>
-            <TableView fx:id="tableView">
+            <TableView fx:id="tableView" prefHeight="400.0" prefWidth="363.0">
                <columns>
                   <TableColumn fx:id="name" minWidth="0.0" prefWidth="84.0" text="Name" />
                   <TableColumn fx:id="price" prefWidth="78.4000244140625" resizable="false" text="Price" />
                   <TableColumn fx:id="quantity" prefWidth="76.79998779296875" text="Quantity" />
-                  <TableColumn fx:id="description" prefWidth="88.79998779296875" text="Description" />
+                  <TableColumn fx:id="description" prefWidth="89.0" text="Description" />
                </columns>
                <HBox.margin>
                   <Insets right="120.0" />
